Initial Load of Files
Includes all jupyter notebooks for Logistic Regression, KNN, GridSearch, and Random Forest.  It also includes two additional jupyter notebooks that include the exploration of the data (graphical and data clean-up).  
A data dictionary explaining each of the variables is included in the depository.
The original dataset  and the clean version used in the ML testing are included within the data folder.